Some refactoring in reading and listening types
- Combine `ReadingQuestion` and `ListeningQuestion` into one type,
  `LRQuestion`
- Add `QuestionComponentInput` to use in all question components
- Move `Answer` and `SetAnswer` types into `readingUtilities`
